+++ CHANGELOG.md
@@ -84,3 +84,4 @@ Semua perubahan yang dilakukan akan didokumentasikan di dalam file ini.
 - Menambah library `tui-pagination` untuk membagi data agar tidak perlu scroll panjang.
 - Menambah`utils/pagination-helper` untuk kustomisasi pagination
 - Menambahkan **pagination** ke `rooms`.
+- Menambahkan **pagination** ke `dashboard`.